export_word/requirements.txt
python-docx==1.1.0
flask==3.0.2
